% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurize_supported_tools.R
\name{futurize_supported_packages}
\alias{futurize_supported_packages}
\alias{futurize_supported_functions}
\title{List packages and functions supporting futurization}
\usage{
futurize_supported_packages()

futurize_supported_functions(package)
}
\arguments{
\item{package}{A package name.}
}
\value{
A character vector of package or function names.
\code{futurize_supported_functions()} produces an error if packages
required by the futurize transpiler are not installed.
}
\description{
List packages and functions supporting futurization
}
\examples{
pkgs <- futurize_supported_packages()
pkgs

if (requireNamespace("future.apply")) {
  fcns <- futurize_supported_functions("base")
  print(fcns)
}

if (requireNamespace("doFuture")) {
  fcns <- futurize_supported_functions("foreach")
  print(fcns)
}

}
