% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpile.R
\name{get_transpiler}
\alias{get_transpiler}
\title{Get a registered transpiler for an R expression}
\usage{
get_transpiler(
  expr,
  envir = parent.frame(),
  unwrap = list(),
  type,
  what,
  debug = FALSE
)
}
\arguments{
\item{expr}{The R expression, which contains an R symbol or an R call,
to be transpiled.}

\item{envir}{The environment where the expression should be evaluated.}

\item{unwrap}{(optional) A list of functions that should be considered
wrapping function, that the transpiler should unwrap ("enter"). This
Allows us to transpile expressions within \code{{ ... }} and \code{local( ... )}.}

\item{type}{Type of the transpiler to use.}

\item{what}{A character string used in error messages describing what
type of transpiler is used.}

\item{debug}{If TRUE, output debug information.}
}
\value{
A transpiler, which is a named list with elements:
\itemize{
\item \code{label} - a character string describing the transpiler
\item \code{transpiler} - a function that takes an R expression and
an optional argument \code{options}
}
}
\description{
Get a registered transpiler for an R expression
}
