% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_call.R
\name{parse_call}
\alias{parse_call}
\title{Parse an R call to be transpiled}
\usage{
parse_call(call, envir = parent.frame(), what = "transpiler", debug = FALSE)
}
\arguments{
\item{call}{An R symbol or an R call.}

\item{envir}{The environment from which to search for the function that
is called in the \code{call}.}

\item{what}{A character string used in error messages describing what
type of transpiler is used.}

\item{debug}{If TRUE, output debug information.}
}
\value{
A names list with elements:
\itemize{
\item \code{fcn}: the function called in the \code{call}
\item \code{fcn_name}: the name of the function
\item \code{ns_name}: the name of the namespace where the function lives
}
}
\description{
Parse an R call to be transpiled
}
\keyword{internal}
