% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\alias{partition.list}
\title{Partition futures into equivalent classes}
\usage{
partition(x, ...)

\method{partition}{list}(x, by = "expr", scheduling = 1, chunk_size = NULL, ...)
}
\arguments{
\item{x}{A container of futures, e.g. a list.}

\item{by}{Named elements of the futures to partition futures by.}

\item{scheduling}{Average number of futures ("chunks") per worker.
If `0.0`, then a single future is used to process all elements
of `X`.
If `1.0` or `TRUE`, then one future per worker is used.
If `2.0`, then each worker will process two futures
(if there are enough elements in `X`).
If `Inf` or `FALSE`, then one future per element of
`X` is used.
Only used if `future.chunk.size` is `NULL`.}

\item{chunk_size}{The average number of elements per future ("chunk").
If `Inf`, then all elements are processed in a single future.
If `NULL`, then argument `future.scheduling` is used.}

\item{\ldots}{Currently not used.}
}
\value{
Returns ...
}
\description{
Partition futures into equivalent classes
}
