% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurize.R
\name{supported_packages}
\alias{supported_packages}
\alias{supported_package_functions}
\title{List packages and functions supporting futurization}
\usage{
supported_packages()

supported_package_functions(package)
}
\arguments{
\item{package}{A package name.}
}
\value{
A character vector of package or function names
}
\description{
List packages and functions supporting futurization
}
\examples{
pkgs <- supported_packages()
pkgs

fcns <- supported_package_functions("base")
fcns

}
