% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpile.R
\name{transpile}
\alias{transpile}
\title{Transpile an R expression}
\usage{
transpile(
  expr,
  options = list(...),
  ...,
  when = TRUE,
  eval = TRUE,
  envir = parent.frame(),
  class,
  flavor = "built-in",
  what = "transpile",
  unwrap = list(base::`{`, base::`(`, base::local, base::I, base::identity),
  debug = FALSE
)
}
\arguments{
\item{expr}{An \R expression, typically a function call to transpile.}

\item{options}{(optional) Named options for the transpilation.}

\item{when}{If TRUE (default), the expression is transpiled, otherwise not.}

\item{eval}{If TRUE (default), the transpiled expression is evaluated,
other it is returned.}

\item{envir}{The environment where the expression should be evaluated.}

\item{flavor}{Flavor of the transpiler to use.}

\item{unwrap}{(optional) A list of functions that should be considered
wrapping function, that the transpiler should unwrap ("enter"). This
Allows us to transpile expressions within \code{{ ... }} and \code{local( ... )}.}
}
\value{
Returns the value of the evaluated expression \code{expr} if \code{eval = TRUE},
otherwise the transpiled expression.
}
\description{
Transpile an R expression
}
\keyword{internal}
