% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurize.R
\name{futurize}
\alias{futurize}
\title{Run a map-reduce call in parallel}
\usage{
futurize(
  expr,
  substitute = TRUE,
  ...,
  stdout = TRUE,
  conditions = "condition",
  flavor = c("addon", "built-in"),
  envir = parent.frame()
)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{stdout}{If TRUE (default), then the standard output is captured,
and re-outputted when \code{value()} is called.
If FALSE, any output is silenced (by sinking it to the null device as
it is outputted).
Using \code{stdout = structure(TRUE, drop = TRUE)} causes the captured
standard output to be dropped from the future object as soon as it has
been relayed. This can help decrease the overall memory consumed by
captured output across futures.
Using \code{stdout = NA} fully avoids intercepting the standard output;
behavior of such unhandled standard output depends on the future backend.}

\item{conditions}{A character string of conditions classes to be captured
and relayed.  The default is to relay all conditions, including messages
and warnings.  To drop all conditions, use \code{conditions = character(0)}.
Errors are always relayed.
Attribute \code{exclude} can be used to ignore specific classes, e.g.
\code{conditions = structure("condition", exclude = "message")} will capture
all \code{condition} classes except those that inherits from the \code{message} class.
Using \code{conditions = structure(..., drop = TRUE)} causes any captured
conditions to be dropped from the future object as soon as it has
been relayed, e.g. by \code{value(f)}. This can help decrease the overall
memory consumed by captured conditions across futures.
Using \code{conditions = NULL} (not recommended) avoids intercepting conditions,
except from errors; behavior of such unhandled conditions depends on the
future backend and the environment from which R runs.}

\item{flavor}{Flavor of futurize transpiler to use.
If `"addon"`, then registered transpilers for packages such as
\pkg{future.apply} and \pkg{furrr} are used.
If `"builtin"`, then built-in transpilers are used.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{\ldots}{Currently not used.}
}
\value{
Returns the value of `call`.
}
\description{
Run a map-reduce call in parallel
}
\examples{
xs <- list(1, 1:2, 1:2, 1:5)
a <- 42

# Sequential call
y <- lapply(X = xs, FUN = function(x, ...) {
  sum(c(x, a), ...)
}, na.rm = TRUE)
   
# Parallelized version
y <- lapply(X = xs, FUN = function(x, ...) {
  sum(c(x, a), ...)
}) |> futurize()
}
