% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurize.R
\name{futurize}
\alias{futurize}
\title{Run a map-reduce call in parallel}
\usage{
futurize(
  expr,
  substitute = TRUE,
  options = futurize_options(...),
  ...,
  flavor = c("addon", "built-in"),
  envir = parent.frame()
)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{options}{Named options controlling how futures are resolved.}

\item{flavor}{Flavor of futurize transpiler to use.
If `"addon"`, then registered transpilers for packages such as
\pkg{future.apply} and \pkg{furrr} are used.
If `"builtin"`, then built-in transpilers are used.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{\ldots}{Names options passed to [future_options()].}
}
\value{
Returns the value of `call`.
}
\description{
Run a map-reduce call in parallel
}
\examples{
xs <- list(1, 1:2, 1:2, 1:5)
a <- 42

# Sequential call
y <- lapply(X = xs, FUN = function(x, ...) {
  sum(c(x, a), ...)
}, na.rm = TRUE)
   
# Parallelized version
y <- lapply(X = xs, FUN = function(x, ...) {
  sum(c(x, a), ...)
}) |> futurize()
}
