% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurize.R
\name{futurize}
\alias{futurize}
\title{Evaluate common R functions in parallel}
\usage{
futurize(
  expr,
  substitute = TRUE,
  options = futurize_options(...),
  ...,
  flavor = c("add-on", "built-in"),
  envir = parent.frame()
)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{options, \ldots}{Named options, passed to \code{\link[=futurize_options]{futurize_options()}},
controlling how futures are resolved.}

\item{flavor}{Flavor of futurize transpiler to use.
If \code{"add-on"}, then registered transpilers for packages such as
\pkg{future.apply} and \pkg{furrr} are used.
If \code{"built-in"}, then built-in transpilers are used.}

\item{envir}{The \link{environment} from where global objects should be
identified.}
}
\value{
Returns the value of the evaluated expression \code{expr}.
}
\description{
\if{html}{\figure{futurize-logo.png}{options: style='float: right;' alt='logo' width='120'}}
}
\examples{
xs <- list(1, 1:2, 1:2, 1:5)

# ------------------------------------------
# Base R apply functions
# ------------------------------------------
# Sequential call
y <- lapply(X = xs, FUN = function(x) {
  sum(x)
})
   
# Parallelized version
y <- lapply(X = xs, FUN = function(x) {
  sum(x)
}) |> futurize()
str(y)

# ------------------------------------------
# purrr map-reduce functions
# ------------------------------------------
if (require("purrr") && requireNamespace("furrr")) {

# Sequential call
y <- map(xs, function(x) {
  sum(x)
})
   
# Parallelized version
y <- map(xs, function(x) {
  sum(x)
}) |> futurize()
str(y)

} ## if (require("purrr") && requireNamespace("furrr"))


# ------------------------------------------
# foreach map-reduce functions
# ------------------------------------------
if (require("foreach") && requireNamespace("doFuture")) {

# Sequential call
y <- foreach(x = xs) \%do\% {
  sum(x)
}
   
# Parallelized version
y <- foreach(x = xs) \%do\% {
  sum(x)
} |> futurize()
str(y)

} ## if (require("foreach") && requireNamespace("doFuture"))


# ------------------------------------------
# foreach map-reduce functions
# ------------------------------------------
if (require("plyr") && requireNamespace("doFuture")) {

# Sequential call
y <- llply(xs, sum)
   
# Parallelized version
y <- llply(xs, sum) |> futurize()
str(y)

} ## if (require("plyr") && requireNamespace("doFuture"))
}
