% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\alias{partition.list}
\title{Partition futures into equivalent classes}
\usage{
partition(x, ...)

\method{partition}{list}(x, by = "expr", scheduling = 1, chunk_size = NULL, ...)
}
\arguments{
\item{x}{A container of futures, e.g. a list.}

\item{by}{Named elements of the futures to partition futures by.}

\item{scheduling}{Average number of futures ("chunks") per worker.
If \code{0.0}, then a single future is used to process all elements
of \code{X}.
If \code{1.0} or \code{TRUE}, then one future per worker is used.
If \code{2.0}, then each worker will process two futures
(if there are enough elements in \code{X}).
If \code{Inf} or \code{FALSE}, then one future per element of
\code{X} is used.
Only used if \code{future.chunk.size} is \code{NULL}.}

\item{chunk_size}{The average number of elements per future ("chunk").
If \code{Inf}, then all elements are processed in a single future.
If \code{NULL}, then argument \code{future.scheduling} is used.}

\item{\ldots}{Currently not used.}
}
\value{
Returns ...
}
\description{
Partition futures into equivalent classes
}
\keyword{internal}
